Rename directory set_roles
